+++ src/handler/inspect.rs
@@ -1,20 +1,24 @@
 use crate::{
     di::{Asyncify, Injectable},
-    from_fn_with_description, Handler, HandlerDescription,
+    from_fn_with_description, Handler, HandlerDescription, HandlerSignature,
 };
 
-use std::sync::Arc;
+use std::{collections::BTreeSet, sync::Arc};
 
 /// Constructs a handler that inspects current state. Like [`map`] but does not
 /// add return value of `f` to the container.
 ///
+/// # Run-time signature
+///
+/// - Obligations: `F::obligations()`
+/// - Outcomes: `BTreeSet::default()`
+///
 /// [`map`]: crate::map
 #[must_use]
 #[track_caller]
-pub fn inspect<'a, F, Input, Output, Args, Descr>(f: F) -> Handler<'a, Input, Output, Descr>
+pub fn inspect<'a, F, Output, Args, Descr>(f: F) -> Handler<'a, Output, Descr>
 where
-    Asyncify<F>: Injectable<Input, (), Args> + Send + Sync + 'a,
-    Input: Send + 'a,
+    Asyncify<F>: Injectable<(), Args> + Send + Sync + 'a,
     Output: 'a,
     Descr: HandlerDescription,
 {
@@ -24,10 +28,9 @@ where
 /// The asynchronous version of [`inspect`].
 #[must_use]
 #[track_caller]
-pub fn inspect_async<'a, F, Input, Output, Args, Descr>(f: F) -> Handler<'a, Input, Output, Descr>
+pub fn inspect_async<'a, F, Output, Args, Descr>(f: F) -> Handler<'a, Output, Descr>
 where
-    F: Injectable<Input, (), Args> + Send + Sync + 'a,
-    Input: Send + 'a,
+    F: Injectable<(), Args> + Send + Sync + 'a,
     Output: 'a,
     Descr: HandlerDescription,
 {
@@ -36,13 +39,13 @@ where
 
 /// [`inspect`] with a custom description.
 #[must_use]
-pub fn inspect_with_description<'a, F, Input, Output, Args, Descr>(
+#[track_caller]
+pub fn inspect_with_description<'a, F, Output, Args, Descr>(
     description: Descr,
     f: F,
-) -> Handler<'a, Input, Output, Descr>
+) -> Handler<'a, Output, Descr>
 where
-    Asyncify<F>: Injectable<Input, (), Args> + Send + Sync + 'a,
-    Input: Send + 'a,
+    Asyncify<F>: Injectable<(), Args> + Send + Sync + 'a,
     Output: 'a,
 {
     inspect_async_with_description(description, Asyncify(f))
@@ -50,28 +53,32 @@ where
 
 /// [`inspect_async`] with a custom description.
 #[must_use]
-pub fn inspect_async_with_description<'a, F, Input, Output, Args, Descr>(
+#[track_caller]
+pub fn inspect_async_with_description<'a, F, Output, Args, Descr>(
     description: Descr,
     f: F,
-) -> Handler<'a, Input, Output, Descr>
+) -> Handler<'a, Output, Descr>
 where
-    F: Injectable<Input, (), Args> + Send + Sync + 'a,
-    Input: Send + 'a,
+    F: Injectable<(), Args> + Send + Sync + 'a,
     Output: 'a,
 {
     let f = Arc::new(f);
 
-    from_fn_with_description(description, move |x, cont| {
-        let f = Arc::clone(&f);
-        async move {
-            {
-                let f = f.inject(&x);
-                f().await;
-            }
+    from_fn_with_description(
+        description,
+        move |x, cont| {
+            let f = Arc::clone(&f);
+            async move {
+                {
+                    let f = f.inject(&x);
+                    f().await;
+                }
 
-            cont(x).await
-        }
-    })
+                cont(x).await
+            }
+        },
+        HandlerSignature::Other { obligations: F::obligations(), outcomes: BTreeSet::default() },
+    )
 }
 
 #[cfg(test)]