+++ src/handler/map.rs
@@ -1,8 +1,9 @@
 use crate::{
-    di::{Asyncify, Injectable, Insert},
-    from_fn_with_description, Handler, HandlerDescription,
+    di::{Asyncify, Injectable},
+    from_fn_with_description, Handler, HandlerDescription, HandlerSignature, Type,
 };
-use std::{ops::ControlFlow, sync::Arc};
+
+use std::{collections::BTreeSet, iter::FromIterator, ops::ControlFlow, sync::Arc};
 
 /// Constructs a handler that passes a value of a new type further.
 ///
@@ -10,18 +11,21 @@ use std::{ops::ControlFlow, sync::Arc};
 /// further in a handler chain.
 ///
 /// See also: [`crate::filter_map`].
+///
+/// # Run-time signature
+///
+/// - Obligations: `Projection::obligations()`
+/// - Outcomes: `BTreeSet::from_iter(vec![Type::of::<NewType>()])`
 #[must_use]
 #[track_caller]
-pub fn map<'a, Projection, Input, Output, NewType, Args, Descr>(
+pub fn map<'a, Projection, Output, NewType, Args, Descr>(
     proj: Projection,
-) -> Handler<'a, Input, Output, Descr>
+) -> Handler<'a, Output, Descr>
 where
-    Input: Clone,
-    Asyncify<Projection>: Injectable<Input, NewType, Args> + Send + Sync + 'a,
-    Input: Insert<NewType> + Send + 'a,
+    Asyncify<Projection>: Injectable<NewType, Args> + Send + Sync + 'a,
     Output: 'a,
     Descr: HandlerDescription,
-    NewType: Send,
+    NewType: Send + Sync + 'static,
 {
     map_with_description(Descr::map(), proj)
 }
@@ -29,69 +33,72 @@ where
 /// The asynchronous version of [`map`].
 #[must_use]
 #[track_caller]
-pub fn map_async<'a, Projection, Input, Output, NewType, Args, Descr>(
+pub fn map_async<'a, Projection, Output, NewType, Args, Descr>(
     proj: Projection,
-) -> Handler<'a, Input, Output, Descr>
+) -> Handler<'a, Output, Descr>
 where
-    Input: Clone,
-    Projection: Injectable<Input, NewType, Args> + Send + Sync + 'a,
-    Input: Insert<NewType> + Send + 'a,
+    Projection: Injectable<NewType, Args> + Send + Sync + 'a,
     Output: 'a,
     Descr: HandlerDescription,
-    NewType: Send,
+    NewType: Send + Sync + 'static,
 {
     map_async_with_description(Descr::map_async(), proj)
 }
 
 /// [`map`] with a custom description.
 #[must_use]
-pub fn map_with_description<'a, Projection, Input, Output, NewType, Args, Descr>(
+#[track_caller]
+pub fn map_with_description<'a, Projection, Output, NewType, Args, Descr>(
     description: Descr,
     proj: Projection,
-) -> Handler<'a, Input, Output, Descr>
+) -> Handler<'a, Output, Descr>
 where
-    Input: Clone,
-    Asyncify<Projection>: Injectable<Input, NewType, Args> + Send + Sync + 'a,
-    Input: Insert<NewType> + Send + 'a,
+    Asyncify<Projection>: Injectable<NewType, Args> + Send + Sync + 'a,
     Output: 'a,
     Descr: HandlerDescription,
-    NewType: Send,
+    NewType: Send + Sync + 'static,
 {
     map_async_with_description(description, Asyncify(proj))
 }
 
 /// [`map_async`] with a custom description.
 #[must_use]
-pub fn map_async_with_description<'a, Projection, Input, Output, NewType, Args, Descr>(
+#[track_caller]
+pub fn map_async_with_description<'a, Projection, Output, NewType, Args, Descr>(
     description: Descr,
     proj: Projection,
-) -> Handler<'a, Input, Output, Descr>
+) -> Handler<'a, Output, Descr>
 where
-    Input: Clone,
-    Projection: Injectable<Input, NewType, Args> + Send + Sync + 'a,
-    Input: Insert<NewType> + Send + 'a,
+    Projection: Injectable<NewType, Args> + Send + Sync + 'a,
     Output: 'a,
     Descr: HandlerDescription,
-    NewType: Send,
+    NewType: Send + Sync + 'static,
 {
     let proj = Arc::new(proj);
 
-    from_fn_with_description(description, move |container: Input, cont| {
-        let proj = Arc::clone(&proj);
+    from_fn_with_description(
+        description,
+        move |container, cont| {
+            let proj = Arc::clone(&proj);
 
-        async move {
-            let proj = proj.inject(&container);
-            let res = proj().await;
-            std::mem::drop(proj);
+            async move {
+                let proj = proj.inject(&container);
+                let res = proj().await;
+                std::mem::drop(proj);
 
-            let mut intermediate = container.clone();
-            intermediate.insert(res);
-            match cont(intermediate).await {
-                ControlFlow::Continue(_) => ControlFlow::Continue(container),
-                ControlFlow::Break(result) => ControlFlow::Break(result),
+                let mut intermediate = container.clone();
+                intermediate.insert(res);
+                match cont(intermediate).await {
+                    ControlFlow::Continue(_) => ControlFlow::Continue(container),
+                    ControlFlow::Break(result) => ControlFlow::Break(result),
+                }
             }
-        }
-    })
+        },
+        HandlerSignature::Other {
+            obligations: Projection::obligations(),
+            outcomes: BTreeSet::from_iter(vec![Type::of::<NewType>()]),
+        },
+    )
 }
 
 #[cfg(test)]